restructures project folders
