+++ src/main/gba_checker.ts
