+++ src/renderer/components/config.scss
