+++ src/renderer/components/lobby.scss
