+++ src/renderer/components/lobby.tsx
