+++ src/renderer/components/setup.tsx
