+++ src/renderer/components/sidebar.scss
