+++ src/renderer/components/sidebar.tsx
