+++ src/renderer/renderer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '@/app/app';
+import App from './app';
 
 ReactDOM.render(<App />, document.getElementById('root'));
